Move tapi PredefinedContentHelper to src/utils/tapi
